Chore: logging util function
Chore: helper function sepration
